refactor(hive_event_bus): split tests.rs (#1266)
tests.rs 489 -> 257 LOC; new tests_extra.rs.
